Refactor JSON editor action container into its own component
