+++ packages/compass-editor/src/index.ts
@@ -1,4 +1,4 @@
-export type { CompletionWithServerInfo } from './types';
+export type { CompletionWithServerInfo, EditorRef } from './types';
 export { prettify } from './prettify';
 export type { FormatOptions } from './prettify';
 export {
@@ -7,14 +7,8 @@ export {
   setCodemirrorEditorValue,
   getCodemirrorEditorValue,
 } from './editor';
-export type {
-  EditorView,
-  Command,
-  Annotation,
-  Action,
-  EditorRef,
-  Completer,
-} from './editor';
+export type { EditorView, Command, Annotation, Completer } from './editor';
+export type { Action } from './action-button';
 export { createDocumentAutocompleter } from './codemirror/document-autocompleter';
 export { createValidationAutocompleter } from './codemirror/validation-autocompleter';
 export { createQueryAutocompleter } from './codemirror/query-autocompleter';
